requirements.txt
streamlit==1.45.1
tensorflow==2.11.0
Pillow
numpy
